chore(macros): move helper functions to utils.rs
